+++ src/CUDA_kernels/contribution_function_kernels.h
@@ -10,10 +10,13 @@
 namespace bear{
 
 
-void contributionFunctionGPU(double* contribution_function_dev,
-                             double* absorption_coeff_device, double* wavenumber_list_device,
-                             std::vector<double>& temperature, std::vector<double>& vertical_grid,
-                             const size_t& nb_spectral_points);
+void contributionFunctionGPU(
+  double* contribution_function_dev,
+  float* absorption_coeff_device, 
+  double* wavenumber_list_device,
+  std::vector<double>& temperature, 
+  std::vector<double>& vertical_grid,
+  const size_t& nb_spectral_points);
 
 }
