+++ src/CUDA_kernels/cross_section_kernels.h
@@ -26,20 +26,30 @@
 namespace bear{
 
 
-void calcCrossSectionsHost(double* cross_sections1, double* cross_sections2, double* cross_sections3, double* cross_sections4,
-                           const double temperature1, const double temperature2,
-                           const double log_pressure1, const double log_pressure2,
-                           const double temperature, const double log_pressure, const double number_density,
-                           const size_t nb_spectral_points, const size_t nb_grid_points, const size_t grid_point,
-                           double* absorption_coeff_device, double* scattering_coeff_device);
+void calcCrossSectionsHost(
+  float* cross_sections1, 
+  float* cross_sections2, 
+  float* cross_sections3, 
+  float* cross_sections4,
+  const double temperature1, 
+  const double temperature2,
+  const double log_pressure1, 
+  const double log_pressure2,
+  const double temperature, 
+  const double log_pressure, 
+  const double number_density,
+  const size_t nb_spectral_points, 
+  const size_t nb_grid_points, 
+  const size_t grid_point,
+  float* absorption_coeff_device, 
+  float* scattering_coeff_device);
 
 
-void calcCIACoefficientsHost(double* cross_sections1, double* cross_sections2,
+void calcCIACoefficientsHost(float* cross_sections1, float* cross_sections2,
                              const double temperature1, const double temperature2,
                              const double temperature, const double number_densities,
                              const size_t nb_spectral_points, const size_t nb_grid_points, const size_t grid_point,
-                             double* absorption_coeff_device);
-
+                             float* absorption_coeff_device);
 
 void calcHmContinuumHost(const double hm_number_density,
                          const double h_number_density,
@@ -47,13 +57,17 @@ void calcHmContinuumHost(const double hm_number_density,
                          const double temperature,
                          const int nb_spectral_points, const int grid_point,
                          double* wavelengths_device,
-                         double* absorption_coeff_device);
+                         float* absorption_coeff_device);
 
 
-void initCrossSectionsHost(const size_t nb_points,  double* absorption_coeff_device);
+void initCrossSectionsHost(
+  const size_t nb_points, 
+  float* absorption_coeff_device);
 
-void checkCrossSectionsHost(const size_t nb_spectral_points, const size_t nb_grid_points,
-                                     double* absorption_coeff_device);
+void checkCrossSectionsHost(
+  const size_t nb_spectral_points, 
+  const size_t nb_grid_points,
+  float* absorption_coeff_device);
 
 }
