+++ src/cloud_model/fixed_cloud_model.h
@@ -49,9 +49,9 @@ class FixedCloudModel: public CloudModel{
       const std::vector<double>& parameters, 
       const Atmosphere& atmosphere,
       SpectralGrid* spectral_grid,
-      double* optical_depth_dev, 
-      double* single_scattering_dev, 
-      double* asym_param);
+      float* optical_depth_dev, 
+      float* single_scattering_dev, 
+      float* asym_param);
     protected:
       std::vector<std::vector<double>> fixed_optical_depth;
       std::vector<std::vector<double>> fixed_single_scattering_albedo;
