+++ src/cloud_model/grey_cloud_model.h
@@ -45,9 +45,9 @@ class GreyCloudModel: public CloudModel{
     virtual void opticalPropertiesGPU(
       const std::vector<double>& parameters, const Atmosphere& atmosphere,
       SpectralGrid* spectral_grid,
-      double* optical_depth_dev, 
-      double* single_scattering_dev, 
-      double* asym_param);
+      float* optical_depth_dev, 
+      float* single_scattering_dev, 
+      float* asym_param);
 
   protected:
     bool fixed_bottom = false;
