+++ src/transport_coeff/transport_coeff.cpp
@@ -270,8 +270,8 @@ void TransportCoefficients::calculateGPU(
   const std::vector<double>& number_densities,
   const size_t nb_grid_points,
   const size_t grid_point,
-  double* absorption_coeff_device,
-  double* scattering_coeff_device)
+  float* absorption_coeff_device,
+  float* scattering_coeff_device)
 {
   for (unsigned int i=0; i<gas_species.size(); i++)
     gas_species[i]->calcTransportCoefficientsGPU(
